+++ HISTORY.md
@@ -1,8 +1,17 @@
 # History
 
+## 0.2.0 – 2023-10-24
+
+* Upgrade python tests & remove python 3.7 [Issue #11](https://github.com/sintel-dev/ml-stars/issues/11) - by @sarahmish
+* Add deprecation warning for old ARIMA primitive [Issue #10](https://github.com/sintel-dev/ml-stars/issues/10) - by @sarahmish
+* Remove python 3.6 [Issue #9](https://github.com/sintel-dev/ml-stars/issues/9) - by @sarahmish
+* Add support for python 3.10 & 3.11 [Issue #8](https://github.com/sintel-dev/ml-stars/issues/8) - by @sarahmish
+
+
 ## 0.1.3 – 2023-09-25
 
 * Update ``keras`` adapter [Issue #7](https://github.com/sintel-dev/ml-stars/issues/7) - by @sarahmish
+
 
 ## 0.1.2 – 2023-08-04